api/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
itsdangerous==2.0.1
marshmallow-sqlalchemy==0.28.1
mysqlclient==2.1.1
numpy==1.23.1
pandas==1.4.3
PyJWT==2.4.0
scipy==1.9.0
sqlalchemy-utils==0.38.3